# Simplify parseStorageClasses