eliminate gotos, replace duplicate code with static_foreach, remove mutable local variable.  Potentially speed things up by (hopefully) matching multiple STCs per loop iteration