@benjones 2019-05-08
The gdc build fails because it doesn't have static foreach, and it appears after a quick grep that there are no "static foreach"s in src.  Is there a good workaround and/or is that likely to change in the future?