@ZombineDev 2019-05-08
As a workaround you can use:
```D
alias Seq(T...) = T;
foreach(name; Seq!("const_", "immutable_", "shared_"))
{ .. }
```