@Basile-z 2019-05-08
In a first time wouldn't the simplification work with non-static loops ? You loose the unrolling but once available for gdc this would be very simple to add the `static` to get the stuff unrolled.