@ibuclaw 2019-05-08
> Is there a good workaround

Not a workaround, but tuples are better fit for what you're trying to do.