@benjones 2019-05-08
I don't think a "dynamic" foreach will work because I'm passing the values to __traits(getMember)?