@benjones 2019-05-08
OK, using @ZombineDev 's tuple hack