@Basile-z 2019-05-08
> I don't think a "dynamic" foreach will work because I'm passing the values to __traits(getMember)?

the array used as ForeachAggregate would had to be a CT one... but NVM since the other trick will work ;)