@WalterBright 2019-05-09
Compiler speed is highly sensitive to lexing/parsing speed. The use of state variables (such as `match`) replacing control flow tends to be slower. It is worthwhile in the lexer/parser to be less concerned with elegance than with every cycle of speed.

The use of switch is almost always faster than a sequence of if statements, because compilers have specialized strategies for generating code for them according to the various patters of case values.

Also, use the form `if (` rather than `if(` to conform to usual practice in DMD sources.

My last objection is while the revised code uses significantly fewer lines, which is good, it's also significantly harder to understand, as it requires a solid understanding of metaprogramming and introspection, while the original code is trivial.