@benjones 2019-05-09
I ran a quick benchmark (are there better ones available somewhere?)
HOST_DMD=generated/osx/release/64/dmd /usr/bin/time ./src/build.d -f -j1  

on master: 6.75s, 6.75s, 6.73s real time
on this PR: 6.76s, 6.71s, 6.71s real time

It would appear that this has negligible impact on speed.

Both versions have basically 1 state variable, stc in the original, matched in mine.

Looking at the codegen, the the switch statement is a bunch of cmp, je pairs, so it doesn't look like the compiler is doing anything clever there.

However, my version is about 1.5x as many instructions, which seems to be due to updateStorageClass not being inlined.  Can I do something to suggest/force inlining?  

Because my if statements are sequential, it's possible that fewer loop iterations would be required overall because we can find multiple classes per iteration if the clauses are ordered well.

WRT to the complexity/difficulty of understanding, I have a few rebuttals comments.  
This is a compiler, so presumably people reading/writing code will have a strong grasp of D and its features/idioms.  If Andrei is going to highlight introspection as one of our greatest strengths, let's use it in our most visible software.  

In terms of readability, I think this version makes it more clear that for each of "static, final, auto, etc", we treat them all the same way.  (having access to static foreach would eliminate a bit more ugliness too).  

I appreciate the review.  If you think there's worthwhile stuff here, let me know and I'll fix the spaces.  New habits are hard :)