@WalterBright 2019-05-09
> It would appear that this has negligible impact on speed.

Benchmarking this by timing the execution of dmd itself on an untargeted benchmark is unlikely to give useful results: the timing is too coarse, far to many other variables affecting speed is at play, and the test examples are likely testing something else. I've seen many people led astray by this. (The ambiguous nature of the numbers you posted illustrate.) A far better way is to instrument the target code with a high precision timer, and craft a test case that tickles just what is under test.

> Both versions have basically 1 state variable, stc in the original, matched in mine.

The `match` variable is what I was referring to. The reason `goto` was used was to avoid state variables.

> Looking at the codegen, the the switch statement is a bunch of cmp, je pairs, so it doesn't look like the compiler is doing anything clever there.

That's highly dependent on what compiler is used, and the number/density of the case values.

> Can I do something to suggest/force inlining?

The -inline switch when compiling.

> let's use it in our most visible software.

The most visible software is Phobos. You're right that the dmd source should be an example of doing things right, but that doesn't imply finding ways to use every feature of D. In this example, the metaprogramming doesn't work very cleanly because it is all broken up into various sections that do things a little differently. The switch, on the other hand, is a simple (and complete) list of all the tokens that can appear next. It's more comprehensible with less effort on the part of the reader.