@benjones 2019-05-09
> Benchmarking this by timing the execution of dmd itself on an untargeted benchmark is unlikely to give useful results: the timing is too coarse, far to many other variables affecting speed is at play, and the test examples are likely testing something else. I've seen many people led astray by this. (The ambiguous nature of the numbers you posted illustrate.) A far better way is to instrument the target code with a high precision timer, and craft a test case that tickles just what is under test.

This was a response to your comment about compiler speed being sensitive to these things, not about microbenchmarking.  Compiling a reasonably sized project seems like a good way to determine whether or not there's a significant performance impact.  I haven't profiled dmd, but I would assume that it doesn't spend a lot of time parsing storage classes.

> > Both versions have basically 1 state variable, stc in the original, matched in mine.
> 
> The `match` variable is what I was referring to. The reason `goto` was used was to avoid state variables.

Not sure I follow?  Both versions have exactly 1 state variable.  

> The -inline switch when compiling.

I assumed that dmd was built with -inline... is it not?

> The most visible software is Phobos. You're right that the dmd source should be an example of doing things right, but that doesn't imply finding ways to use every feature of D. In this example, the metaprogramming doesn't work very cleanly because it is all broken up into various sections that do things a little differently. The switch, on the other hand, is a simple (and complete) list of all the tokens that can appear next. It's more comprehensible with less effort on the part of the reader.

I would say "D's metaprogramming allowed me to remove duplicated code and express a generic operation in a way that would be impossible in many languages".  It's clear we disagree on whether that improves the code.  I'll close the PR